sa-div-announcement-worker/requirements.txt
beautifulsoup4==4.6.0
pymongo==3.6.0
PyYAML==3.12
requests==2.18.4
